requirements.txt
# Core FastAPI and async dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0

numpy

# HTTP requests and web scraping
httpx==0.26.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml>=5.0.0

# Database and caching
redis==5.0.1

# AI/ML and NLP
google-generativeai==0.3.2

# Email services
python-dotenv==1.0.0

# Background tasks
celery==5.3.4

# Environment management
dnspython

selenium
razorpay

email-validator
bcrypt
pyjwt

prisma